requirements.txt
pyfunctional~=1.4.3
